moves bodyzoneselector out of components
