+++ tgui/packages/tgui/interfaces/common/BodyZoneSelector.tsx
@@ -1,7 +1,7 @@
 import { Component, createRef } from 'react';
 
-import { resolveAsset } from '../assets';
-import { Image } from './Image';
+import { resolveAsset } from '../../assets';
+import { Image } from '../../components/Image';
 
 export enum BodyZone {
   Head = 'head',
@@ -15,7 +15,7 @@ export enum BodyZone {
   Groin = 'groin',
 }
 
-const bodyZonePixelToZone = (x: number, y: number): BodyZone | null => {
+function bodyZonePixelToZone(x: number, y: number): BodyZone | null {
   // TypeScript translation of /atom/movable/screen/zone_sel/proc/get_zone_at
   if (y < 1) {
     return null;
@@ -52,7 +52,7 @@ const bodyZonePixelToZone = (x: number, y: number): BodyZone | null => {
   }
 
   return null;
-};
+}
 
 type BodyZoneSelectorProps = {
   onClick?: (zone: BodyZone) => void;
